Land ownership model, lexical revision, lint fix
- Ownership backlog #14/#16 (ADR-0036): owned/mut T forms, owned
  name prefix for params/fields, alloc.destroy paired teardown,
  field-assign move, join/detach handle consumption, value-type
  owned diagnostics, main(owned args: *mut Vec<String>)
- Lexical revision ADR-0037 (D8-D15): local var inference,
  &&/|| as logical ops (and/or/tree keywords removed), nested
  block comments, char literals as Unicode codepoints, prefix
  literal lex errors; stage1/examples/corpus migrated
- lint L004: segment-exact abbreviation matching, id/utf8
  false positives removed; regression tests
- Docs: ADR-0037, SPEC syntax modules, README features,
  CONTEXT.md terms; f32/f16 deferred to backlog #5

+++ tag1/hc/src/codegen/bytecode/decode.rs
@@ -461,7 +461,7 @@ fn decode_pattern(r: &mut Reader) -> Result<IrPattern, String> {
         2 => IrPattern::Int(r.i128()?),
         3 => IrPattern::Float(r.f64()?),
         4 => IrPattern::Str(r.str()?),
-        5 => IrPattern::Char(r.u8()?),
+        5 => IrPattern::Char(r.u32()?),
         _ => return Err(format!("未知 switch 模式 tag {tag}")),
     })
 }
